"Applied fix from trunk framework for revision: 1848263  " ------------------------------------------------------------------------ r1848263 | jleroux | 2018-12-06 08:24:13 +0100 (jeu. 06 déc. 2018) | 13 lignes
Fixed: Error on createShoppingListItem when adding item to cart as anonymous
(OFBIZ-5157)

It can be reproduced by setting the product_store field autoSaveCart to 'Y' 
(In Store 9000 for the demo-data). Then the error occurs upon the first visit of 
the ecommerce-site, provided the user is not logged in.

    First Visit: no userLogin
    OFBiz wants to autosave list
    saving list requires userLogin (forsecuritycheck)
    result: error

Thanks: Mirko Vogelsmeier for report, Benjamin Jugl for the fix
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1848267 13f79535-47bb-0310-9956-ffa450edef68