"Applied fix from trunk framework for revision: 1862278" ------------------------------------------------------------------------ r1862278 | jleroux | 2019-06-28 13:53:55 +0200 (ven. 28 juin 2019) | 17 lignes
Fixed: User should be notified with success message on MRP run in manufacturing component
(OFBIZ-9533)

Steps to regenerate:
1. Go to Manufacturing component (manufacturing/control/main)
2. Click on Mrp sub menu (manufacturing/control/FindInventoryEventPlan)
3. Click on Run Mrp button.(manufacturing/control/RunMrp)
4. Select Facility/Facility Group and Click Submit button.
On success, the user should be notified with success message like 
"Mrp run is scheduled".

jleroux: despite being a sub-task of an improvement I decided to backport. 
It works well and nothing could go wrong, only UI changes.

Thanks: Aditya Sharma for report, Humera Khan for the fix, Prachi Shastri for
testing and Pierre Smits for noticing this could be backported (despite being
a sub-task of an improvement)
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1862281 13f79535-47bb-0310-9956-ffa450edef68